Allow users to select mobile plant permits
https://eaflood.atlassian.net/browse/WE-842

The purpose of this PR is to allow users to select the five types of mobile permit instead of the existing SR2015 No 18.

The task list has been amended to match new format so there are now only 2 sections rather than the previous 3.

Users can now select one of the 5 mobile permits. The existing SR2015 No 18 permit has been hidden and is no longer selectable.

In addition, the initial version of the Invoicing Details page has been added so that there is somewhere to go to when the 'Give invoicing details' task list item is clicked on (as per the other task list items).