Remove red border around file upload control
This change improves the appearance of the file upload controls if a validation error occurs, e.g. a user hasn't selected a file before trying to submit the page.

It was agreed with the designers that a red border shouldn't be added to the control because it didn't have a border beforehand. This is in keeping with other controls in the application such as radio buttons.