Application timeout
https://eaflood.atlassian.net/browse/WE-880

This change adds an expiry date to the session cookie, configured to 3 hours in the .env file. If the cookie expiry is exceeded, requests will forward to a new application timeout page.

To accomplish this and to facilitate a further change that will stop users accessing pages in their application flow that should not be accessible because of the state of their application the validateCookie() method has been refined to return a more granular result rather than a boolean.

This meant that all validateCookie unit tests had to be amended, so a general test helper has been introduced that prevents the same unit test code appearing in multiple route tests.