Prevent users editing application after submission
https://eaflood.atlassian.net/browse/WE-918

This change adds the following new screens:

Already Submitted
Not Complete
Not Paid
Not Submitted
Start at Beginning

The controllers have been amended to redirect to one of these screens if the application is not in the correct state to view the requested route.

In order to achieve this there was some work required on calculating application completeness. This is done in multiple ways and could do with further refactoring at some point.

There is currently no way to mark an application as submitted through the front end, this will need to be done in Dynamics to test the Not Submitted screen.

Also included is some refactoring of the unit tests including:

Standardise use of the GeneralTestHelper.checkElementsExist method in the tests
Some move of stubs to use Sinon sandbox (still more to do on this).