Added 'isPaid' check to all required routes
Added 'isPaid' check to all required routes so that the user is redirected to the 'You need to pay for your application' screen if they have submitted their application (i.e. checked their answers) but have not gone through a payment route.

Also turned off caching to fix a problem of the user pressing the back button and circumventing the page protection.