Back link on 'How do you want to Pay' displays 'Sorry we cannot find that application'
https://eaflood.atlassian.net/browse/WE-1384
https://eaflood.atlassian.net/browse/WE-1309

This fix prevents the following from happening:

"Back link on 'How do you want to Pay' displays 'Sorry we cannot find that application' and so does recovering from the link in the save and return email once the 'How do you want to Pay' page has been reached."

The back button now works as intended however in the case of "Grabbing the save application link" from the email, the user will be returned to the tasklist (with all tasks displayed as complete) and the user need only to click "Send application and pay" to return to the "Check your answers page".

This was accomplished by not setting the "submitted on" date when the tasklist has been completed but instead after a successful card payment is made or when using Bacs, after the confirmation of using Bacs is confirmed.

This means that checking if a payment has been made is no longer necessary and has made the processing much simpler. Effectively once the user has started their application but have not yet sent it, they are able to press back or jump to any page if they have it book-marked except for the "Application received" page. If they have sent it, they will no longer be able to press back or jump to any book-marked page except for the "Application received" page. Instead they will see the "Application has already been sent" page.

All the unit tests no longer need to mock the payment entity and this processing has also been removed. Please note that an earlier update to the "standard" package forced me to make sure all JSON objects are formatted correctly with the correct spacing between the curly brackets.