More route unit tests now use standard mocks
Also fixes https://eaflood.atlassian.net/browse/WE-2047 as follows:
The issue was found in the recovery service when a controller (e.g. the check your emails controller).

Fixed by making sure that when the applicationID cookie is not found, the context should not attempt to contain the recovered application details.

All affected unit tests have been changed so they are using the standard mocking library including correctly mocking the recovery service.