Ignore no-deprecated-api rule
Standard warns us that the `url.parse` method has been deprecated since Node 11 and should be replaced with the `url.URL` constructor. However this is not a drop-in replacement, and would require a rewrite of code which is best left for another branch. We therefore ignore this rule for the time being.