Replace gulp-uglify with gulp-terser
uglify, which is run as part of the build process, fails because it doesn't support ES6 features, some of which we now use as part of the Standard fixes. terser is a fork of uglify which is updated for modern features. We therefore replace uglify with it so our CI will pass.