Update Node, fix linting and get unit tests passing in CI (#857)
The first step of updating this project is to get the tests passing in CI. Although they previously passed, it turns out this should not have been the case! Running standard from the shell found multiple linting errors, but these were not being flagged up as part of the CI. Investigation found that gulp-standard was simply not raising the errors, thus allowing CI to pass! Updating gulp-standard resolved this.

This change therefore updates gulp-standard to the latest version, along with standard, and fixes the various linting errors. All linting errors are fully fixed, with the exception of a deprecated API. Standard warns us that the url.parse method has been deprecated since Node 11 and should be replaced with the url.URL constructor. However this is not a drop-in replacement, and would require a larger rewrite of code which is best left for another branch. We therefore ignore this rule for the time being.

We replace gulp-uglify with gulp-terser; uglify, part of our build process, now fails because the Standard fixes use some ES6 features, which uglify does not support. We therefore drop in terser, a modern fork of uglify, which is updated for these features.

Finally, we update Node to v14, as the updated packages will not run in v10.

These steps get the unit tests passing in CI. They do not get CI fully passing, as SonarCloud now re-analyzes the modified files and has many things to say about them. But this will be addressed in a future PR.

* Update standard and gulp-standard

We update standard to the latest version, v16. We also update gulp-standard as the previous version was passing even if linting errors were found! v14 resolves this.

* Automated fixes

This is the initial batch of fixes done by running `standard --fix`

* Resolve multi-line ternary errors in cli templates

* Use `const` instead of `let` for variables which aren't reassigned

* Resolve lexical declarations in case block

Standard wil raise an error if variables are declared in a case block and the block isn't wrapped in {} brackets

* Additional ternary expression fixes

* Fix Object.prototype hasOwnProperty method errors

Standard does not allow Object.prototype methods to be accessed from an object; Object.prototype.method should be called instead. In this case, we have statements such as:

request.payload.hasOwnProperty(treatment.questionCode)

To be Standard-compliant, we rewrite them as:

Object.prototype.hasOwnProperty.call(request.payload, treatment.questionCode)

* Ignore no-deprecated-api rule

Standard warns us that the `url.parse` method has been deprecated since Node 11 and should be replaced with the `url.URL` constructor. However this is not a drop-in replacement, and would require a rewrite of code which is best left for another branch. We therefore ignore this rule for the time being.

* Add globals to be ignored

When running tests, we are warned that leaks are detected in FinalizationRegistry and WeakRef. Since the aim of this change is to fix linting and get the CI passing, we ignore these globals for now.

* Update node to 14

The updated packages will fail with node 10. We therefore update node in `package.json` and `ci.yml` to 14 to ensure everything works.

* Replace gulp-uglify with gulp-terser

uglify, which is run as part of the build process, fails because it doesn't support ES6 features, some of which we now use as part of the Standard fixes. terser is a fork of uglify which is updated for modern features. We therefore replace uglify with it so our CI will pass.

* Update `node-sass`

As observed by @bensagar-ea, `node-sass` fails during CI as version 4.13.1 does not support node 14. To resolve this we bump it to 4.14.1, which does support node 14.

* Revert linter changes for client-side JS

The file src/assets/javascripts/application.js is run client-side; to ensure compatibility, we revert the changes imposed by Standard and add a linting exception to this file to exclude it from Standard's checks.

* Add additional paths to `gulp standard` job

Previously, the `gulp standard` job would only lint files in the `src/` folder. This change adds in `cli/`, `test/` plus any .js files in the top-level folder.

* First fix of SonarCloud issues

Some of the SonarCloud errors require further discussion to agree how they are to be resolved. However many of them (eg. line length) can be quickly resolved so they are implemented in this change.

* Set up for npm 6

The work was done in a dev environment with npm 7. However we have been advised by webops to use npm 6, which is not guaranteed to be compatible with the output of npm 7. This change deletes the npm 7 file in favour of an npm 6 file.

* Add SonarCloud properties file

Adding sonar-project.properties allows us to specify which folders to check. This means we can exclude the `test/` folder which is needlessly having a negative impact our on quality report!