First pass at setting up govuk-frontend
We start by setting up the gulp build task to copy and install the govuk-frontend assets; amending application.scss to import the files; and initialising the javascript in bodyEnd.html. This allows the example component on the first page to work correctly. The styling is not quite correct yet but this could be due to clashes with the existing stylesheets.