Exclude imported frontend javascript from linting
Our branch is failing CI due to linting errors. This is caused by the govuk-frontend javascript, which is copied to the src/assets/javascripts/govuk/ folder during the gulp build task, which occurs prior to linting. As we're unable to fix these linting errors, we update the gulp standard task to exclude javascript files in this folder.