Add default validation compiler to `baseRoute.js`
As of v19 Hapi no longer provides a default validation compiler. To fix this we update `baseRoute.js` to pass the validation payload to Joi.compile(), which will then compile an uncompiled schema or pass through an already-compiled schema (which is the case with validators which return a Joi object, eg. NeedToConsultValidator)