Rewrite tests to correctly pass/fail
The tests as written would only execute the assertions in the catch block if an error was thrown; so if no error was thrown then the tests would skip over the entire catch block and therefore pass instead of fail. This change rewrites the tests to ensure the the assertions are executed regardless of whether or not an error is thrown.