Apply deny list to exclude certain routes
The tests for bespokeApplyOffline check the routes to see which ones have it as the controller. It then iterates over each of these routes and runs the tests. However, 2 of the 3 routes with this controller don't have behaviour defined within the route. This causes the tests to fail. Since we only want to update the dependencies at present and not change the service's behaviour, we therefore add a deny list to exclude the failing routes. If and when the controller is updated for these routes, they can be removed from the list.