Update hapi to v20 (#862)
Updating hapi to v20 introduces a couple of breaking changes, which this PR fixes:

Hapi no longer provides a default validation compiler, so we register joi as our default validator in server.js;
options.payload.multipart now defaults to false instead of true, breaking the upload facility, so we therefore explicitly set it to true.

* Update hapi to v20

* Add default validation compiler to `baseRoute.js`

As of v19 Hapi no longer provides a default validation compiler. To fix this we update `baseRoute.js` to pass the validation payload to Joi.compile(), which will then compile an uncompiled schema or pass through an already-compiled schema (which is the case with validators which return a Joi object, eg. NeedToConsultValidator)

* Fix upload route errors

As of Hapi v19, `options.payload.multipart` for routes defaults to `false` instead of `true`, breaking the upload facility. We therefore explicitly set it to `true` to get the upload facility working again.

* Set default validator in more standard way

Instead of compiling validation in `baseRoute.js` we instead register `joi` as our default validator in `server.js`