Update nock to v13 (#864)
Updating dev dependency nock to v13 causes one of the tests to break due to it using a deprecated method socketDelay. This change fixes the test:
- Replaces socketDelay with the method delayConnection;
- increases the delayConnection timeout past the configured timeout value, which causes nock to send a timeout response immediately.
- While making this change, we discovered that the tests expecting an error to be thrown would pass even if no error was thrown; this was due to the assertions being wrapped in a catch block, so if no error was thrown then the assertions were skipped over and the test would therefore pass. We therefore rewrite the tests so that the assertions are always executed, regardless of whether or not an error is thrown.

* Update nock to v13

* Fix test using deprecated method

* Rewrite tests to correctly pass/fail

The tests as written would only execute the assertions in the catch block if an error was thrown; so if no error was thrown then the tests would skip over the entire catch block and therefore pass instead of fail. This change rewrites the tests to ensure the the assertions are executed regardless of whether or not an error is thrown.