Update code to v8 (#865)
Updating the code dev dependency causes the bespokeApplyOffline tests to fail for three routes. On further investigation, it isn't clear why these were passing in the first place as the to.only.include() assertion was previously passing when comparing two unequal lists of items, which should not have been the case!

The first issue is that the mocks are not being set up correctly. Changing how we set up the waste activities mock gets the test passing for one of the three routes, but the other two still fail.

This is due to the controller only assigning the on-page items if the route's itemType is wasteActivity. This means that the facility route fails the tests as itemType is not set, and the waste assessment route fails as itemType is wasteAssessment. We therefore need to change how the tests work.

We refactor the original bespokeApplyOffline test to only apply to one route, not all three. We rename it wasteActivityApplyOffline accordingly, and create separate test files facilityApplyOffline and wasteAssessmentApplyOffline which do most of the tests of the original, but confirming that the list of assessments is empty (which is the current behaviour).

* Update code to v8

* Fix test mocks

* Apply deny list to exclude certain routes

The tests for bespokeApplyOffline check the routes to see which ones have it as the controller. It then iterates over each of these routes and runs the tests. However, 2 of the 3 routes with this controller don't have behaviour defined within the route. This causes the tests to fail. Since we only want to update the dependencies at present and not change the service's behaviour, we therefore add a deny list to exclude the failing routes. If and when the controller is updated for these routes, they can be removed from the list.

* Change `bestpokeApplyOffline` test to `wasteActivityApplyOffline` test

The original test was generic and applied to all routes with the `bespokeApplyOffline` controller. Here we rework it to apply only to waste activities and be more explicit about what it tests

* Small test refactor

* Add additional route tests

Now that we have decided on one test file per route, we add additional `facilityApplyOffline` and `wasteAssessmentApplyOffline` tests