Update gulp-sass to v5
gulp-sass v5.0.0 no longer includers a compiler so we install sass as a dev dependency.

Note that the updated version of sass gives multiple warnings due to  govuk_frontend_toolkit and govuk-elements-sass using sass features which are deprecated and which will not be available in sass v2. Therefore it may not be possible to update sass for much longer unless either govuk_frontend_toolkit and govuk-elements-sass are updated, or this service is updated to solely use govuk-design-system