Update gulp-sass to v5 (#868)
Updating gulp-sass to v5 gives us build errors as it no longer includes a compiler. We therefore install sass as a dev dependecy. We also remove node-sass as it has been deprecated, and we have no need for it now that we've installed sass.

Note that the updated version of sass gives multiple warnings due to govuk_frontend_toolkit and govuk-elements-sass using sass features which are deprecated and which will not be available in sass v2. Therefore it may not be possible to update sass for much longer unless either govuk_frontend_toolkit and govuk-elements-sass are updated, or this service is updated to solely use govuk-design-system.

* Update gulp-sass to v5

gulp-sass v5.0.0 no longer includers a compiler so we install sass as a dev dependency.

Note that the updated version of sass gives multiple warnings due to  govuk_frontend_toolkit and govuk-elements-sass using sass features which are deprecated and which will not be available in sass v2. Therefore it may not be possible to update sass for much longer unless either govuk_frontend_toolkit and govuk-elements-sass are updated, or this service is updated to solely use govuk-design-system

* Remove node-sass

node-sass has been deprecated in favour of sass; and as we have just installed sass for use as a compiler we no longer need node-sass and can therefore remove it as a dependency

* Make more readable

gulp-sass's suggested way of using it, `const sass = require('gulp-sass')(require('sass'))`, was not felt to be particularly readable. We therefore break it up into separate assignments to be more comprehensible at a glance