Install and set up govuk-frontend package (#861)
After updating the service's dependencies, we now aim to install govuk-frontend which will allow us to use its cookie banner component in future. This change does the following:
-Sets up the gulp build task to copy and install the govuk-frontend assets;
- amends application.scss to import the files, and sets govuk-frontend to compatibility mode which will correctly apply its styling when installed alongside the existing frontend toolkit;
- initialises the javascript in bodyEnd.html;
- prevents javascript files in the src/assets/javascripts/govuk/ folder from being linted by the gulp standard task to stop CI failing on files we can't change.

After testing this with the accordion and cookie banner components, we can confirm that this allows us to drop the JavaScript versions of components into our views and they will respond correctly.

* Install `govuk-frontend` package

* Add example component to first page of service

* First pass at setting up govuk-frontend

We start by setting up the gulp build task to copy and install the govuk-frontend assets; amending application.scss to import the files; and initialising the javascript in bodyEnd.html. This allows the example component on the first page to work correctly. The styling is not quite correct yet but this could be due to clashes with the existing stylesheets.

* Remove example component from first page

* Create cookieBanner.html and add to first page

We create a partial cookieBanner.html and add it to the first page. It looks like there isn't a template we can add this banner to; therefore it may be necessary to add it to every page.

* Fix styling with govuk-frontend compatibility mode

govuk-frontend components will be incorrectly styled if used alongside the older govuk frontend toolkits unless we enable its compatibility mode, which we do in this change.

* Remove cookieBanner.html

The cookieBanner.html partial was only introduced as a way of checking that the new govuk-frontend will work alongisde the existing one; having demonstrated this can be done, we can now remove cookieBanner.html

* Exclude imported frontend javascript from linting

Our branch is failing CI due to linting errors. This is caused by the govuk-frontend javascript, which is copied to the src/assets/javascripts/govuk/ folder during the gulp build task, which occurs prior to linting. As we're unable to fix these linting errors, we update the gulp standard task to exclude javascript files in this folder.